Fix rms_norm backward precision issue for weight grad
- Change torch.sum to use float32 for accumulation before converting to input dtype
- This fixes precision loss when summing partial_buffer in weight grad calculation
- Before: max relative error up to 772% for large shapes (100, 4096)
- After: max relative error < 1% for all tested shapes

The issue was caused by converting to float16 during sum operation,
which accumulates rounding errors. Now we sum in float32 first,
then convert to the target dtype, preserving precision.

+++ src/flag_gems/ops/rms_norm.py
@@ -213,7 +213,7 @@ class RmsNorm(torch.autograd.Function):
                 ROW_BLOCK_SIZE,
                 COL_BLOCK_SIZE,
             )
-            dw = torch.sum(partial_buffer, dim=0, dtype=x.dtype).reshape(-1)
+            dw = torch.sum(partial_buffer, dim=0, dtype=torch.float32).to(x.dtype).reshape(-1)
 
         return dx, None, dw, None
